backend/requirements.txt
sqlmodel
fastapi[standard]
pyjwt
pwdlib[argon2]
